Subchannel Test - Add support for creating subchannels - Modify memory access calls - Move Libfabric-specific functions away from trf.h/trf.c - Remove cursor data update functions - Drop C standard to 99 for better Windows compatibility (future release) - Allow users to send/receive raw data in addition to delimited   messages, both on the TCP channel and fabric channel

+++ libtrf/src/pbdef.sh
@@ -1,4 +1,15 @@
+if [ ! -f /usr/bin/protod-c ]
+then
+    echo "Protocol Buffers C compiler required."
+    echo "Please install it from your distribution's package manager."
+    echo "Fedora, RHEL, and derivatives: protobuf-c-devel protobuf-c-compiler"
+    echo "Debian, Ubuntu, and derivatives: libprotobuf-c-dev protobuf-c-compiler"
+    exit 1
+fi
+
 rm -f trf_msg.pb-c.c ../include/trf_msg.pb-c.h
 protoc-c --c_out=. trf_msg.proto
 sed -i 's,\/\*$,/**,g' trf_msg.pb-c.h
 mv trf_msg.pb-c.h ../include/trf_msg.pb-c.h
+
+echo "Regenerated trf_msg.pb-c.c and trf_msg.pb-c.h"
